+++ packages/mint-components/CHANGELOG.md
@@ -14,7 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed Component:
   - /<sqm-leaderboard>
     - Now supports a top point earners leaderboard, `topPointEarners`
-    - MAx number of leaderboard rows displayed is now configurable via prop but defaults to 10
+    - Added a `programId` prop to overwride program context
+      - If an empty string is passed, then a global leaderboard is displayed
+    - Max number of leaderboard rows displayed is now configurable via prop but defaults to 10
 
 ## [1.5.3] - 2022-03-31